# backend/openapi.json
{
  "components" : {
    "schemas" : {
...
      "Pageable" : {
        "properties" : {
          "page" : {
            "format" : "int32",
            "minimum" : 0,
            "type" : "integer"
          },
          "size" : {
            "format" : "int32",
            "minimum" : 1,
            "type" : "integer"
          },
          "sort" : {
            "items" : {
              "type" : "string"
            },
            "type" : "array"
          }
        },
        "type" : "object"
      },
...
      "PagedResponseUserSummary" : {
        "properties" : {
          "items" : {
            "items" : {
              "$ref" : "#/components/schemas/UserSummary"
            },
            "type" : "array"
          },
          "page" : {
            "format" : "int32",
            "type" : "integer"
          },
          "size" : {
            "format" : "int32",
            "type" : "integer"
          },
          "totalElements" : {
            "format" : "int64",
            "type" : "integer"
          },
          "totalPages" : {
            "format" : "int32",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
...
      "UpdateEnabledRequest" : {
        "properties" : {
          "enabled" : {
            "type" : "boolean"
          }
        },
        "required" : [ "enabled" ],
        "type" : "object"
      },
...
      "UpdateProfileRequest" : {
        "properties" : {
          "age" : {
            "format" : "int32",
            "maximum" : 120,
            "minimum" : 13,
            "type" : "integer"
          },
          "city" : {
            "type" : "string"
          },
          "displayName" : {
            "type" : "string"
          },
          "firstName" : {
            "type" : "string"
          },
          "lastName" : {
            "type" : "string"
          },
          "locale" : {
            "type" : "string"
          },
          "notificationPreference" : {
            "enum" : [ "EMAIL", "PUSH", "NONE" ],
            "type" : "string"
          },
          "phone" : {
            "pattern" : ".*\\S.*",
            "type" : "string"
          },
          "timezone" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "UpdateRoleRequest" : {
        "properties" : {
          "role" : {
            "enum" : [ "USER", "ADMIN" ],
            "type" : "string"
          }
        },
        "required" : [ "role" ],
        "type" : "object"
      },
      "UserDetail" : {
        "properties" : {
          "accountType" : {
            "type" : "string"
          },
          "age" : {
            "format" : "int32",
            "type" : "integer"
          },
          "city" : {
            "type" : "string"
          },
          "createdAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "displayName" : {
            "type" : "string"
          },
          "email" : {
            "type" : "string"
          },
          "enabled" : {
            "type" : "boolean"
          },
          "firstName" : {
            "type" : "string"
          },
          "id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "lastName" : {
            "type" : "string"
          },
          "locale" : {
            "type" : "string"
          },
          "notificationPreference" : {
            "enum" : [ "EMAIL", "PUSH", "NONE" ],
            "type" : "string"
          },
          "phone" : {
            "type" : "string"
          },
          "role" : {
            "type" : "string"
          },
          "timezone" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "UserSummary" : {
        "properties" : {
          "accountType" : {
            "type" : "string"
          },
          "createdAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "displayName" : {
            "type" : "string"
          },
          "email" : {
            "type" : "string"
          },
          "enabled" : {
            "type" : "boolean"
          },
          "id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "role" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
    }
  },
  "info" : {
    "description" : "Generated API documentation for the OpenSkeleton backend.",
    "license" : {
      "name" : "MIT"
    },
    "title" : "OpenSkeleton API",
    "version" : "v1"
  },
  "openapi" : "3.1.0",
  "paths" : {
...
    "/api/v1/admin/users" : {
      "get" : {
        "operationId" : "listUsers",
        "parameters" : [ {
          "in" : "query",
          "name" : "pageable",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/Pageable"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PagedResponseUserSummary"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "admin-user-controller" ]
      }
    },
    "/api/v1/admin/users/{id}" : {
      "get" : {
        "operationId" : "getUser",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/UserDetail"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "admin-user-controller" ]
      }
    },
    "/api/v1/admin/users/{id}/enabled" : {
      "patch" : {
        "operationId" : "updateEnabled",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UpdateEnabledRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/UserSummary"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "admin-user-controller" ]
      }
    },
    "/api/v1/admin/users/{id}/profile" : {
      "patch" : {
        "operationId" : "updateProfile",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UpdateProfileRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/UserDetail"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "admin-user-controller" ]
      }
    },
    "/api/v1/admin/users/{id}/role" : {
      "patch" : {
        "operationId" : "updateRole",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UpdateRoleRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/UserSummary"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "admin-user-controller" ]
      }
    },
...
  },
  "servers" : [ {
    "description" : "Generated server url",
    "url" : "http://localhost"
  } ]
}